requirements.txt
Flask==2.0.3
pytest==7.0.1
pytest-cov==3.0.0
requests==2.27.1
Flask-SQLAlchemy==2.5.1
flask-marshmallow==0.14.0
marshmallow==3.15.0
marshmallow-sqlalchemy==0.28.0
